+++ CHANGES.md
@@ -10,11 +10,14 @@ Release Notes.
 - Index: Remove sortable field from the stored field. If a field is sortable only, it won't be stored.
 - Index: Support InsertIfAbsent functionality which ensures documents are only inserted if their docIDs are not already present in the current index. There is a exception for the documents with extra index fields more than the entity's index fields.
 - Measure: Introduce "index_mode" to save data exclusively in the series index, ideal for non-timeseries measures.
+- Index: Use numeric index type to support Int and Float
 
 ### Bug Fixes
 
 - Fix the bug that TopN processing item leak. The item can not be updated but as a new item.
 - Resolve data race in Stats methods of the inverted index.
+- Fix the bug when adding new tags or fields to the measure, the querying crashes or returns wrong results.
+- Fix the bug that adding new tags to the stream, the querying crashes or returns wrong results.
 
 ### Documentation